+++ Frameworks/ShortcutRecorder/Source/SRRecorderControl.h
@@ -31,7 +31,7 @@
 #pragma mark *** Key Combination Control ***
 
 - (unsigned int)allowedFlags;
-- (void)setAllowedFlags:(unsigned int)flags;
+- (void)setAllowedFlags:(NSUInteger)flags;
 
 - (BOOL)allowsKeyOnly;
 - (void)setAllowsKeyOnly:(BOOL)nAllowsKeyOnly escapeKeysRecord:(BOOL)nEscapeKeysRecord;
@@ -41,7 +41,7 @@
 - (void)setCanCaptureGlobalHotKeys:(BOOL)inState;
 
 - (unsigned int)requiredFlags;
-- (void)setRequiredFlags:(unsigned int)flags;
+- (void)setRequiredFlags:(NSUInteger)flags;
 
 - (KeyCombo)keyCombo;
 - (void)setKeyCombo:(KeyCombo)aKeyCombo;
@@ -61,13 +61,13 @@
 
 #pragma mark *** Conversion Methods ***
 
-- (unsigned int)cocoaToCarbonFlags:(unsigned int)cocoaFlags;
-- (unsigned int)carbonToCocoaFlags:(unsigned int)carbonFlags;
+- (NSUInteger)cocoaToCarbonFlags:(NSUInteger)cocoaFlags;
+- (NSUInteger)carbonToCocoaFlags:(NSUInteger)carbonFlags;
 
 @end
 
 // Delegate Methods
 @interface NSObject (SRRecorderDelegate)
-- (BOOL)shortcutRecorder:(SRRecorderControl *)aRecorder isKeyCode:(signed short)keyCode andFlagsTaken:(unsigned int)flags reason:(NSString **)aReason;
+- (BOOL)shortcutRecorder:(SRRecorderControl *)aRecorder isKeyCode:(signed short)keyCode andFlagsTaken:(NSUInteger)flags reason:(NSString **)aReason;
 - (void)shortcutRecorder:(SRRecorderControl *)aRecorder keyComboDidChange:(KeyCombo)newKeyCombo;
 @end
