+++ Plugins/Twitter Plugin/MGTwitterEngine/MGTwitterEngine.h
@@ -55,7 +55,7 @@
 - (void)setClearsCookies:(BOOL)flag;
 
 // Connection methods
-- (int)numberOfConnections;
+- (NSUInteger)numberOfConnections;
 - (NSArray *)connectionIdentifiers;
 - (void)closeConnection:(NSString *)identifier;
 - (void)closeAllConnections;
